app/pyproject.toml
[tool.poetry]
name = "multi-agent-backend"
version = "0.0.0"
description = ""
authors = ["Aman Sharma <aman.sharma562993@gmail.com>"]
readme = "README.md"
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
flask = ">=3.1.1,<4.0.0"
requests = ">=2.32.3,<3.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
sqlalchemy = ">=2.0.41,<3.0.0"
flask-sqlalchemy = ">=3.1.1,<4.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
pytest = ">=8.4.0,<9.0.0"
pyjwt = "^2.10.1"
sqlalchemy-utils = "^0.41.2"
flask-migrate = "^4.1.0"
langchain = "^0.3.26"
langchain-google-genai = "^2.1.5"
langchain-google-community = "^2.0.7"
langchain-community = "^0.3.26"
google-search-results = "^2.4.2"
langgraph = "^0.4.8"
langgraph-supervisor = "^0.0.27"
langchain-tavily = "^0.2.4"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
mypy = "^1.16.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
